🐛 Add message check to toggle 'allow-empty'
Provides a work around for git aborting a commit due to the commit
message being the same as the tempalte even though all the details
may have been provided.

If a message arg is provided then we assume no other changes to the commit
message need to be made so we enable the `--allow-empty-message` flag.

This allows us to commit even though there is no diff between the commit msg
& the template.